add secu admin

+++ src/Controller/CheckHasLikeController.php
@@ -6,7 +6,6 @@ use App\Entity\Like;
 use App\Repository\LikeRepository;
 use App\Repository\RecipeRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use phpDocumentor\Reflection\Types\Integer;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -16,21 +15,26 @@ final class CheckHasLikeController extends AbstractController
     #[Route('/check/has/like/add/{id}', name: 'app_check_has_like_add')]
     public function add(int $id, EntityManagerInterface $em, RecipeRepository $recipeRepository): Response
     {
+        $user = $this->getUser();
         $recipe = $recipeRepository->findOneBy(['id' => $id]);
         if(!$recipe) {
             throw $this->createNotFoundException(
                 'Recette introuvable à l\'id : '. $id
             );
         }
-        $like = new Like();
+        if($user) {
+            $like = new Like();
 
-        $like->setRecipe($recipe);
-        $like->setUser($this->getUser());
-        $like->setCreatedAt(new \DateTime());
-        $em->persist($like);
-        $em->flush();
-        $this->addFlash('success', 'Like ajouté !');
-        return $this->redirectToRoute('app_recipe', ['id' => $id]);
+            $like->setRecipe($recipe);
+            $like->setUser($this->getUser());
+            $like->setCreatedAt(new \DateTime());
+            $em->persist($like);
+            $em->flush();
+            $this->addFlash('success', 'Like ajouté !');
+            return $this->redirectToRoute('app_recipe', ['id' => $id]);
+        } else {
+            return $this->redirectToRoute('app_login');
+        }
     }
 
     #[Route('/check/has/like/remove/{id}', name: 'app_check_has_like_remove')]
@@ -42,11 +46,13 @@ final class CheckHasLikeController extends AbstractController
             throw $this->createNotFoundException(
                 'Recette introuvable avec l\'id : '.$id
             );
-        } else {
+        } elseif($user) {
             $em->remove($like);
             $em->flush();
             $this->addFlash('success', 'Like supprimer !');
             return $this->redirectToRoute('app_recipe', ['id' => $id]);
+        } else {
+            return $this->redirectToRoute('app_login');
         }
     }
 }
